backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
click==8.3.3
fastapi==0.136.1
h11==0.16.0
idna==3.13
numpy==2.4.4
opencv-contrib-python-headless==4.13.0.92
pillow==12.2.0
pydantic==2.13.4
pydantic_core==2.46.4
python-multipart==0.0.27
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.46.0
